requirements.txt
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
Markdown==2.6.9
olefile==0.44
Pillow==4.2.1
pkg-resources==0.0.0
pytz==2017.2
